Allow ability to enable/disable blocks with groups
https://github.com/ben-wallis/Filtration/issues/40

For any given block that uses block groups, those blocks will be
enabled/disabled rather than toggled between show/hide if the following
syntax is used:

[show|hide] #! [group_spec]

Notice the additional "!" immediately after the "#" and before the first
group name.

It is possible to have some blocks which toggle Enable/Disable and some
which toggle Show/Hide for a given group.